[Refactor] : TCA 0.55.1 -> 1.0.0
Git : #43
