+++ Projects/UI/AD_UI/Sources/Views/MakeADViews/Step2_FindingTheCharacterView/CropImageView/CropImageView.swift
