+++ Projects/UI/AD_UI/Sources/Views/MakeADViews/Step3_SeparatingCharacterView/MaskingImageView/Helper/TouchDownPanGestureRecognizer.swift
